Pipes.Internal

Description

This is an internal module, meaning that it is unsafe to import unless you understand the risks.

This module provides a fast implementation by weakening the monad transformer laws. These laws do not hold if you can pattern match on the constructors, as the following counter-example illustrates:

lift . return = M . return . Pure

return = Pure

lift . return /= return

You do not need to worry about this if you do not import this module, since the other modules in this library do not export the constructors or export any functions which can violate the monad transformer laws.

Internal

data Proxy a' a b' b m r

A Proxy is a monad transformer that receives and sends information on both an upstream and downstream interface.

The type variables signify:

a' and a - The upstream interface, where (a')s go out and (a)s come in

b' and b - The downstream interface, where (b)s go out and (b')s come in

m - The base monad

r - The return value

Constructors

Request a' (a -> Proxy a' a b' b m r)

Respond b (b' -> Proxy a' a b' b m r)

M (m (Proxy a' a b' b m r))

Pure r

unsafeHoist :: Monad m => (forall x. m x -> n x) -> Proxy a' a b' b m r -> Proxy a' a b' b n r

unsafeHoist is like hoist, but faster.

This is labeled as unsafe because you will break the monad transformer laws if you do not pass a monad morphism as the first argument. This function is safe if you pass a monad morphism as the first argument.

observe :: Monad m => Proxy a' a b' b m r -> Proxy a' a b' b m r

The monad transformer laws are correct when viewed through the observe function:

observe (lift (return r)) = observe (return r)

observe (lift (m >>= f)) = observe (lift m >>= lift . f)

This correctness comes at a small cost to performance, so use this function sparingly.

This function is a convenience for low-level pipes implementers. You do not need to use observe if you stick to the safe API.

type X = Void

The empty type, used to close output ends

closed :: X -> a

Use closed to "handle" impossible outputs